+++ src/main/java/com/example/demo/repository/impl/IntentRepositoryImpl.java
@@ -76,7 +76,7 @@ public class IntentRepositoryImpl implements IntentRepository {
 
     @Override
     @Transactional
-    public Integer getMaxResult() {
+    public float getMaxResult() {
         return maxResultConfidence.getMaxResult();
     }
